fix(v1.5): clippy lints (nix build -D clippy::all gate)
Der nix-build läuft clippy mit -D clippy::all (strenger als cargo test/build) und
brach an 3 v1.5-Lints ab:
- absence.rs:743/795 (Phase 19): manuelles x = x + Duration → x += Duration (assign_op_pattern)
- reporting.rs:984 (Phase 22): unnötiger Cast current_week as u8 → current_week (unnecessary_cast)

Verifiziert: cargo clippy --workspace -- -D clippy::all grün. (Pre-existing
Doc-Overindent-Lints in test/reporting_no_contract_volunteer.rs sind NICHT aus v1.5
und blockieren den nix-Lib-Build nicht.)

+++ service_impl/src/reporting.rs
@@ -981,7 +981,7 @@ impl<Deps: ReportingServiceDeps> service::reporting::ReportingService
             .get_report_for_employee(
                 sales_person_id,
                 year,
-                current_week as u8,
+                current_week,
                 context.clone(),
                 tx,
             )
